chore: relocate roadmap websocket service file
